details_microservice/requirements.txt
Flask
requests
redis
python-dotenv
